+++ app/index.tsx
@@ -4,7 +4,7 @@ import { useFonts } from 'expo-font';
 import { useColorScheme } from 'react-native';
 
 import { StatusBar } from 'expo-status-bar';
-import { Main } from '../src/Main';
+import { Main } from './main';
 import themes from '../src/themes';
 import ThemeContextProvider from '../src/contexts/ThemeContext';
